Report os and browser distinctly
Keep darwin as the os name since we are giving darwin version as the os
version. That way we can map it to a mac version in the ui consistently.
In the same vein, report playwright-{browser} so that maybe we can map
the correction browser version in the ui.