Project cleanup (#16)
* Project cleanup

* Convert eslintrc to .js from .json
* Use a standard tsconfig base
* Update packages

* fix ts errors

* Remove saucelabs package dependency

* Cleanup deps again

* Forgot to add the base ts config
* Remove babel

* fix up ts errors

* Fix attachments

* Add suffix for attachments
* Create readable streams properly

* clean up unneeded type

* Report os and browser distinctly

Keep darwin as the os name since we are giving darwin version as the os
version. That way we can map it to a mac version in the ui consistently.
In the same vein, report playwright-{browser} so that maybe we can map
the correction browser version in the ui.